// src/app/core/interfaces/response.interface.ts
export interface IResponse{
    success: boolean;
    data: {
        id: string;
    nombre: string;
    apellido: string;
    cedula: string;
    clave: string;
    email: string;
    email_p: string;
    notificar: string;
    telefono: string;
    celular: string;
    direccion: string;
    f_nacimiento: string;
    departamento: string;
    puesto: string;
    salario: string;
    token: string;
    };
}